fix: :bug: Ensure mortality for cancer is larger than the general population
Previously, the mortality for people with and without cancer were assessed separately, and mortality for people with cancer was independent of age. This lead to some weird results, where mortality for some older ages could be greater if a person did not have cancer, than if it had it. This commit fixes this bug by considering that cancer mortality is always additional to a baseline (non-cancer related) age-adjusted mortality.